+++ trixnity-messenger/src/commonMain/kotlin/de/connect2x/trixnity/messenger/MatrixMessengerConfiguration.kt
@@ -3,8 +3,9 @@ package de.connect2x.trixnity.messenger
 import de.connect2x.trixnity.client.MatrixClientConfiguration
 import de.connect2x.trixnity.client.ModuleFactory
 import de.connect2x.trixnity.messenger.util.mb
-import io.ktor.client.*
-import io.ktor.client.engine.*
+import io.ktor.client.HttpClient
+import io.ktor.client.HttpClientConfig
+import io.ktor.client.engine.HttpClientEngine
 import kotlin.time.Duration
 import kotlin.time.Duration.Companion.hours
 
@@ -47,6 +48,7 @@ data class MatrixMessengerConfiguration(
     val features: Features = Features(
         enablePdfReader = true,
         enablePowerlevelEventConfigurationInRoomSettings = true,
+        enableMessageDrafts = true
     ),
 
     /**
@@ -154,6 +156,7 @@ data class MatrixMessengerConfiguration(
          */
         var enablePdfReader: Boolean = true,
         var enablePowerlevelEventConfigurationInRoomSettings: Boolean = true,
+        var enableMessageDrafts: Boolean = true
     )
 
     enum class CryptoDriver {