+++ trixnity-messenger/src/commonTest/kotlin/de/connect2x/trixnity/messenger/viewmodel/roomlist/RoomListViewModelTest.kt
@@ -261,6 +261,9 @@ class RoomListViewModelTest {
             )
         } returns flowOf(null)
 
+        every { roomServiceMock.getDraftMessage(any()) } returns flowOf(null)
+        every { roomServiceMock2.getDraftMessage(any()) } returns flowOf(null)
+
         every { onRoomSelectedMock.invoke(any(), any()) } returns Unit
 
         every { onAccountSelected.invoke() } returns Unit
@@ -1184,7 +1187,8 @@ class RoomListViewModelTest {
                                     override val accounts: StateFlow<List<AccountInfo>> =
                                         MutableStateFlow(listOf())
                                     override val globalNotificationCount: StateFlow<String?> = MutableStateFlow(null)
-                                    override val accountNotificationCounts: StateFlow<Map<UserId, String?>> = MutableStateFlow(emptyMap())
+                                    override val accountNotificationCounts: StateFlow<Map<UserId, String?>> =
+                                        MutableStateFlow(emptyMap())
 
                                     override fun selectActiveAccount(userId: UserId?) {
                                         backgroundScope.launch {